✂️ Separates setup-server responsabilities into other modules
