+++ src/server/schemas.ts
@@ -1,0 +1,33 @@
+const UserRequestBodySchema = {
+  schema: {
+    body: {
+      type: "object",
+      properties: {
+        name: {
+          type: "string",
+        },
+        email: {
+          type: "string",
+        },
+        password: {
+          type: "string",
+          minLength: 6,
+          pattern: "(?=.*[A-Za-z])(?=.*\\d)",
+        },
+        birthDate: {
+          type: "string",
+        },
+      },
+      required: ["name", "email", "password", "birthDate"],
+    },
+  },
+};
+
+interface RequestBody {
+  name: string;
+  email: string;
+  password: string;
+  birthDate: string;
+}
+
+export { RequestBody, UserRequestBodySchema };
